feat(core): Moved DynamicBaseMemberType to it's own file
