+++ Core/Core/Models/DynamicBase.cs
@@ -253,30 +253,6 @@ namespace Speckle.Core.Models
   }
 
   /// <summary>
-  /// Represents all different types of members that can be returned by <see cref="DynamicBase.GetMembers"/>
-  /// </summary>
-  [Flags]
-  public enum DynamicBaseMemberType
-  {
-    /// <summary>
-    /// The typed members of the DynamicBase object
-    /// </summary>
-    Instance = 1,
-    /// <summary>
-    /// The dynamically added members of the DynamicBase object
-    /// </summary>
-    Dynamic = 2,
-    /// <summary>
-    /// The typed members flagged with <see cref="ObsoleteAttribute"/> attribute.
-    /// </summary>
-    Obsolete = 4,
-    /// <summary>
-    /// The typed members flagged with <see cref="SchemaIgnored"/> attribute.
-    /// </summary>
-    SchemaIgnored = 8,
-  }
-  
-  /// <summary>
   /// This attribute is used internally to hide the this[key]{get; set;} property from inner reflection on members.
   /// For more info see this discussion: https://speckle.community/t/why-do-i-keep-forgetting-base-objects-cant-use-item-as-a-dynamic-member/3246/5
   /// </summary>
